requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0.1
Flask>=3.0.0
Markdown>=3.5.0
python-dateutil>=2.8.2
jinja2>=3.1.2
